+++ include/fitterbap/comm/timesync.h
@@ -24,6 +24,7 @@
 #define FBP_COMM_TIMESYNC_H_
 
 #include "fitterbap/cmacro_inc.h"
+#include "fitterbap/time.h"
 #include <stdint.h>
 #include <stdbool.h>
 
@@ -33,9 +34,12 @@
  *
  * @brief A simple time synchronization protocol.
  *
- * This module implements a simple time synchronization method.
+ * This module implements a simple time synchronization method using
+ * the client's fbp_time_counter() and the server-provide UTC time.
  *
- *
+ * The server is always considered as time source truth.  If an
+ * implementation uses multiple servers, then only one server
+ * connection should be connected to update UTC time.
  *
  * ## References
  *
@@ -47,51 +51,69 @@
 
 FBP_CPP_GUARD_START
 
+
+/// Opaque instance
+struct fbp_ts_s;
+
+
 /**
- * @brief Get the platform-specific counter value.
+ * @brief The current local, monotonic counter value.
  *
- * @return The counter value, which rolls over.
- *
- * The counter frequency should be at least 1000 Hz (1 ms period).
- * Higher frequency allows for increased accuracy, both for time
- * convergence and individual samples.  Most applications should
- * have a frequency around 1,000,000 Hz (1 MHz).
+ * @return The counter value.
  */
-FBP_API uint32_t fbp_ts_platform_counter();
+static inline uint64_t fbp_ts_counter() {
+    return fbp_time_counter().value;
+}
 
 /**
  * @brief Get the current time.
  *
+ * @param The timesync instance.  NULL will use the time provided by
+ *      the first created instance, if it exists.  This behavior enables a
+ *      simple singleton interface, which is common for most applications.
  * @return The current time.  The format is 34Q30 as defined by
- *      fitterbap/time.h.
+ *      fitterbap/time.h.  This function returns 0 if the UTC
+ *      time is not yet known.
+ * @see fbp_time_utc
  *
  * When using this module to provide time, edit your fbp/config.h file
- * to have:
- * static inline int64_t fbp_time_get() { return fbp_ts_time(); }
+ * to define FBP_TS_PROVIDE_UTC.
+ *
+ * This function is thread-safe.
  */
-FBP_API int64_t fbp_ts_time();
+FBP_API int64_t fbp_ts_time(struct fbp_ts_s * self);
 
 /**
- * @brief Update the current time with new data.
+ * @brief Update the timesync instance with new measurement data.
  *
- * @param src_tx The time we sent the TIMESYNC request message.
- * @param tgt_rx The time that the target received the TIMESYNC request message.
- * @param tgt_tx The time that the target send the TIMESYNC response message.
- *
- * During normal operation, call this function at least 1/2 of the
- * platform counter rollover duration.
+ * @param self The instance, or NULL to ignore.
+ * @param src_tx The time we sent the TIMESYNC request message, in
+ *      fbp_ts_counter() ticks.
+ * @param tgt_rx The time that the target received the TIMESYNC request
+ *      message, in the server's UTC time.
+ * @param tgt_tx The time that the target send the TIMESYNC response message,
+ *      in the server's UTC time.
+ * @param src_rx The time we received the TIMESYNC request message, in
+ *      fbp_ts_counter() ticks.
  */
-FBP_API void fbp_ts_update(int64_t src_tx, int64_t tgt_rx, int64_t tgt_tx);
+FBP_API void fbp_ts_update(struct fbp_ts_s * self, uint64_t src_tx, int64_t tgt_rx, int64_t tgt_tx, uint64_t src_rx);
 
 /**
- * @brief Update the time using the platform counter.
+ * @brief Allocate and initialize a new instance.
  *
- * During normal operation, call fbp_ts_update() at least 1/2 of the
- * platform counter rollover duration.  When this device is disconnected,
- * the system cannot call fbp_ts_update().  Call this function
- * instead.
+ * @return The new instance.
  */
-FBP_API void fbp_ts_process();
+FBP_API struct fbp_ts_s * fbp_ts_initialize();
+
+/**
+ * @brief Finalize and free an existing instance.
+ *
+ * @param self The instance created by fbp_ts_initialize().
+ *
+ * Most embedded systems should never use this function.
+ */
+FBP_API void fbp_ts_finalize(struct fbp_ts_s * self);
+
 
 FBP_CPP_GUARD_END