+++ src/host/win/uart.c
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#define FBP_LOG_LEVEL FBP_LOG_LEVEL_DEBUG1
+#define FBP_LOG_LEVEL FBP_LOG_LEVEL_NOTICE
 
 #include "fitterbap/host/uart.h"
 #include "fitterbap/cdef.h"